% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_p.R
\name{calculate_p}
\alias{calculate_p}
\title{Calculating p function}
\usage{
calculate_p(xbeta, ff, epsilon, outcome)
}
\arguments{
\item{xbeta}{my fitted betas matrix multiplied by covariate}

\item{ff}{my estimated f's}

\item{epsilon}{probability of observing a gene when it should be absent; probability between 0 and 1}

\item{outcome}{length-n vector; this is the vector of a target gene's presence/absence; should be coded as 0 or 1}
}
\value{
a vector of probabilities
}
\description{
Calculating p function
}
\examples{
data("TM7_data")
outcome <- TM7_data$`Ribosomal protein L27`
# create design matrix
covariate <- model.matrix(~tongue, data = TM7_data) 
beta <- c(0.1,0.03)
my_xbeta <- covariate \%*\% beta
nn <- length(outcome)
my_ff <- rep(mean(outcome), nn)
epsilon <- 0 

my_calculated_p <- calculate_p(xbeta = my_xbeta, ff = my_ff, epsilon = 0, outcome = outcome)
}
