% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{expit: inverse of logit}
\usage{
expit(input)
}
\arguments{
\item{input}{input a vector of value(s). These values can be between negative infinity to infinity.}
}
\value{
Returns value(s) that are proportion(s) between 0 and 1
}
\description{
expit: inverse of logit
}
\examples{
# taking the expit of any value(s) between negative infinity and infinity & return a proportion 
expit(3)
# taking the expit of a vector of values 
expit(c(-0.1, 2, 1.5, -1))
}
