% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInits_f.R
\name{genInits_f}
\alias{genInits_f}
\title{Function for generating initial starts for estimated f}
\usage{
genInits_f(nn, nstarts = 1, seed, outcome)
}
\arguments{
\item{nn}{length of outcome nn}

\item{nstarts}{number of starts}

\item{seed}{numeric seed for random initializations}

\item{outcome}{vector of outcome, binary 0 & 1}
}
\value{
matrix of initializations
}
\description{
Function for generating initial starts for estimated f
}
\examples{
outcome <- rbinom(10, 1,.5)
genInits_f(nn = 10, nstarts = 1, seed = 88, outcome = outcome)
}
