% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_happi.R
\name{happi}
\alias{happi}
\title{Main function for p=q=1}
\usage{
happi(
  outcome,
  covariate,
  quality_var,
  max_iterations = 50,
  h0_param = 2,
  nstarts = 1,
  change_threshold = 0.05,
  epsilon = 0
)
}
\arguments{
\item{outcome}{length-n vector}

\item{covariate}{n x p matrix}

\item{quality_var}{length-n vector  TODO(turn into n x q matrix)}

\item{max_iterations}{the maximum number of EM steps that the algorithm will run for}

\item{h0_param}{the column index in covariate that has beta=zero under the null}

\item{nstarts}{number of starts TODO}

\item{change_threshold}{algorithm will terminate early if the likelihood changes by this percentage or less for 5 iterations in a row for both the alternative and the null}

\item{epsilon}{prob of ...}
}
\description{
Main function for p=q=1
}
