% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_happi_test.R
\name{happi_test}
\alias{happi_test}
\title{Testing function running happi, p=q=1; this script contains the modularized version of happi with correct implementation of log likelihood}
\usage{
happi_test(
  outcome,
  covariate,
  quality_var,
  max_iterations = 1000,
  min_iterations = 15,
  h0_param = 2,
  change_threshold = 0.1,
  epsilon = 0,
  method = "splines",
  random_starts = FALSE,
  firth = TRUE,
  spline_df = 3,
  nstarts = 1,
  seed = 13
)
}
\arguments{
\item{outcome}{length-n vector; this is the vector of a target gene's presence/absence; should be coded as 0 or 1}

\item{covariate}{n x p matrix; this is the matrix for the primary predictor/covariate of interest}

\item{quality_var}{length-n vector; this is the quality variable vector, currently p = 1  TODO(turn into n x q matrix)}

\item{max_iterations}{the maximum number of EM steps that the algorithm will run for}

\item{min_iterations}{the minimum number of EM steps that the algorithm will run for}

\item{h0_param}{the column index in covariate that has beta=zero under the null}

\item{change_threshold}{aalgorithm will terminate early if the likelihood changes by this percentage or less for 5 iterations in a row for both the alternative and the null}

\item{epsilon}{probability of observing a gene when it should be absent; probability between 0 and 1}

\item{method}{method for estimating f. Defaults to "splines" which fits a monotone spline with df determined by 
argument spline_df; "isotone" for isotonic regression fit}

\item{random_starts}{whether to pick the starting values of beta's randomly. Defaults to FALSE.}

\item{firth}{use firth penalty? Default is TRUE.}

\item{spline_df}{degrees of freedom (in addition to intercept) to use in
monotone spline fit; default 3}

\item{nstarts}{number of starts; Integer. Defaults to \code{1}. Number of starts for optimization.}

\item{seed}{numeric number to set seed for random multiple starts}
}
\value{
An object of class \code{happi}.
}
\description{
Testing function running happi, p=q=1; this script contains the modularized version of happi with correct implementation of log likelihood
}
