% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npLRT.R
\name{npLRT}
\alias{npLRT}
\title{Non parametric LRT aka permutation test}
\usage{
npLRT(
  happi_out,
  max_iterations = 1000,
  min_iterations = 15,
  h0_param = 2,
  change_threshold,
  epsilon,
  method,
  firth,
  spline_df,
  nstarts,
  P = 1000
)
}
\arguments{
\item{happi_out}{a happi results object}

\item{max_iterations}{the maximum number of EM steps that the algorithm will run for}

\item{min_iterations}{the minimum number of EM steps that the algorithm will run for}

\item{h0_param}{the column index in covariate that has beta=zero under the null}

\item{change_threshold}{algorithm will terminate early if the likelihood changes
by this percentage or less for 5 iterations in a row for both the alternative and the null}

\item{epsilon}{probability of observing a gene when it should be absent; probability between 0 and 1}

\item{method}{method for estimating f. Defaults to "splines" which fits a monotone spline with df determined by 
argument spline_df; "isotone" for isotonic regression fit}

\item{firth}{use firth penalty? Default is TRUE.}

\item{spline_df}{degrees of freedom (in addition to intercept) to use in
monotone spline fit; default 3}

\item{nstarts}{number of starts; Integer. Defaults to \code{1}. Number of starts for optimization.}

\item{P}{number of permutation iterations}
}
\value{
An object with npLRT pvalues
}
\description{
Non parametric LRT aka permutation test
}
