% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbLRT.R
\name{pbLRT}
\alias{pbLRT}
\title{parametric boostrap LRT implementation}
\usage{
pbLRT(
  happi_out,
  max_iterations = 3000,
  min_iterations = 15,
  h0_param = 2,
  change_threshold = 0.1,
  epsilon = 0,
  method = "splines",
  firth = T,
  spline_df = 4,
  nstarts = 1,
  B = 1000
)
}
\arguments{
\item{happi_out}{a happi results object}

\item{max_iterations}{the maximum number of EM steps that the algorithm will run for}

\item{min_iterations}{the minimum number of EM steps that the algorithm will run for}

\item{h0_param}{the column index in covariate that has beta=zero under the null}

\item{change_threshold}{algorithm will terminate early if the likelihood changes
by this percentage or less for 5 iterations in a row for both the alternative and the null}

\item{epsilon}{probability of observing a gene when it should be absent; probability between 0 and 1}

\item{method}{method for estimating f. Defaults to "splines" which fits a monotone spline with df determined by 
argument spline_df; "isotone" for isotonic regression fit}

\item{firth}{use firth penalty? Default is TRUE.}

\item{spline_df}{degrees of freedom (in addition to intercept) to use in
monotone spline fit; default 3}

\item{nstarts}{number of starts; Integer. Defaults to \code{1}. Number of starts for optimization.}

\item{B}{number of bootstrap iterations}
}
\value{
An object with pbLRT pvalues
}
\description{
parametric boostrap LRT implementation
}
