% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_em.R
\name{run_em}
\alias{run_em}
\title{General E-M Algorithm}
\usage{
run_em(
  outcome = outcome,
  quality_var = quality_var,
  change_threshold = change_threshold,
  max_iterations = max_iterations,
  min_iterations = min_iterations,
  epsilon = epsilon,
  method = method,
  firth = firth,
  spline_df = spline_df,
  nn = nn,
  em_covariate = NULL,
  em_estimates = NULL,
  em_estimated_beta = NULL,
  em_estimated_basis_weights = NULL,
  em_estimated_ftilde = NULL,
  em_estimated_p = NULL,
  em_fitted_xbeta = NULL,
  em_estimated_f = NULL
)
}
\arguments{
\item{outcome}{length-n vector; this is the vector of a target gene's presence/absence; should be coded as 0 or 1}

\item{quality_var}{length-n vector; this is the quality variable vector, currently p = 1  TODO(turn into n x q matrix)}

\item{change_threshold}{algorithm will terminate early if the likelihood changes by this percentage or less for 5 iterations in a row for both th}

\item{max_iterations}{the maximum number of EM steps that the algorithm will run for}

\item{min_iterations}{the minimum number of EM steps that the algorithm will run for}

\item{epsilon}{probability of observing a gene when it should be absent; probability between 0 and 1}

\item{method}{method for estimating f. Defaults to "splines" which fits a monotone spline with df determined by 
argument spline_df; "isotone" for isotonic regression fit}

\item{firth}{use firth penalty? Default is TRUE.}

\item{spline_df}{degrees of freedom (in addition to intercept) to use in
monotone spline fit; default 3}

\item{nn}{length(outcome)}

\item{em_covariate}{n x p matrix; this is the matrix for the primary predictor/covariate of interest}

\item{em_estimates}{log likelihood estimates}

\item{em_estimated_beta}{estimated betas}

\item{em_estimated_basis_weights}{estimated basis weights}

\item{em_estimated_ftilde}{estimated f_tilde aka logit(estimated_f)}

\item{em_estimated_p}{estimated probablities}

\item{em_fitted_xbeta}{fitted betas}

\item{em_estimated_f}{estimated f's}
}
\value{
An object of class \code{happi}.
}
\description{
General E-M Algorithm
}
