% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wasserman.R
\name{run_wasserman}
\alias{run_wasserman}
\title{# Wasserman split test function}
\usage{
run_wasserman(
  happi_results_object,
  firth = TRUE,
  spline_df = 3,
  max_iterations = 1000,
  change_threshold = 0.01,
  epsilon = 0,
  nstarts = 1,
  h0_param = 2
)
}
\arguments{
\item{happi_results_object}{An object of class \code{happi}.}

\item{firth}{use firth penalty? Default is TRUE.}

\item{spline_df}{degrees of freedom (in addition to intercept) to use in
monotone spline fit; default 3}

\item{max_iterations}{the maximum number of EM steps that the algorithm will run for}

\item{change_threshold}{algorithm will terminate early if the likelihood changes
by this percentage or less for 5 iterations in a row for both the alternative and the null}

\item{epsilon}{probability of observing a gene when it should be absent; probability between 0 and 1}

\item{nstarts}{number of starts; Integer. Defaults to \code{1}. Number of starts for optimization.}

\item{h0_param}{the column index in covariate that has beta=zero under the null}
}
\value{
A list of Wasserman pvalues
}
\description{
# Wasserman split test function
}
