% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_b}
\alias{simulate_b}
\title{Internal simulate function for parametric bootstrap}
\usage{
simulate_b(happi_results_sims, h0_param = 2)
}
\arguments{
\item{happi_results_sims}{an object of happi results used for simulating new data}

\item{h0_param}{the column index in covariate that has beta=zero under the null}
}
\value{
a simulated dataframe using estimated null parameters
}
\description{
Internal simulate function for parametric bootstrap
}
