% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beta.R
\name{update_beta}
\alias{update_beta}
\title{Function for updating betas}
\usage{
update_beta(probs, covariate, firth = TRUE)
}
\arguments{
\item{probs}{outcomes p; probability between 0 and 1}

\item{covariate}{n x p matrix; this is the matrix for the primary predictor/covariate of interest}

\item{firth}{use firth penalty? Default is TRUE.}
}
\value{
vector of estimated betas
}
\description{
Function for updating betas
}
\examples{
data("TM7_data")
# create design matrix
covariate <- model.matrix(~tongue, data = TM7_data) 
outcome <- TM7_data$`Ribosomal protein L27`
nn <- length(outcome)
probs <- rep(0.5,nn)
my_beta <- update_beta(probs = probs, covariate = covariate, firth = TRUE)
}
