% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warningcheck_update_f.R
\name{warningcheck_update_f}
\alias{warningcheck_update_f}
\title{Function for checking update_f}
\usage{
warningcheck_update_f(
  probs,
  tuning_param = 50,
  method = "splines",
  spline_df = spline_df,
  quality_var = quality_var,
  nn = nn,
  outcome = outcome,
  tt = tt
)
}
\arguments{
\item{probs}{outcomes p; probability between 0 and 1}

\item{tuning_param}{tuning parameter. default set to  50}

\item{method}{method for updating f. Can specify either  isotone or splines. default splines.}

\item{spline_df}{degrees of freedom for splines. default 3.}

\item{quality_var}{length-n vector; this is the quality variable vector, currently p = 1}

\item{nn}{length of outcome vector}

\item{outcome}{length-n vector; this is the vector of a target gene's presence/absence; should be coded as 0 or 1}

\item{tt}{E-M iteration value}
}
\value{
warning for checking f update
}
\description{
Function for checking update_f
}
