% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_happi.R
\name{happi}
\alias{happi}
\title{Main function running happi, p=q=1}
\usage{
happi(
  outcome,
  covariate,
  quality_var,
  max_iterations = 50,
  min_iterations = 15,
  h0_param = 2,
  nstarts = 1,
  change_threshold = 0.05,
  epsilon = 0,
  method = "splines",
  random_starts = FALSE,
  firth = TRUE,
  spline_df = 3
)
}
\arguments{
\item{outcome}{length-n vector; this is the vector of a target gene's presence/absence; should be coded as 0 or 1}

\item{covariate}{n x p matrix; this is the matrix for the primary predictor/covariate of interest}

\item{quality_var}{length-n vector; this is the quality variable vector, currently p = 1  TODO(turn into n x q matrix)}

\item{max_iterations}{the maximum number of EM steps that the algorithm will run for}

\item{min_iterations}{the minimum number of EM steps that the algorithm will run for}

\item{h0_param}{the column index in covariate that has beta=zero under the null}

\item{nstarts}{number of starts TODO}

\item{change_threshold}{algorithm will terminate early if the likelihood changes by this percentage or less for 5 iterations in a row for both the alternative and the null}

\item{epsilon}{prob of ...}

\item{method}{method for estimating f. Defaults to "splines" which fits a monotone spline with df determined by 
argument spline_df; "isotone" for isotonic regression fit}

\item{random_starts}{whether to pick the starting values of beta's randomly. Defaults to FALSE.}

\item{firth}{use firth penalty? Default is TRUE.}

\item{spline_df}{degrees of freedom (in addition to intercept) to use in
monotone spline fit; default 3}
}
\description{
Main function running happi, p=q=1
}
\examples{
load(TM7_data)
x_matrix <- model.matrix(~tongue, data = TM7_data) # create design matrix

happi_results <- happi(outcome=TM7_data$`Ribosomal protein L27`, 
covariate=x_matrix, 
quality_var=TM7_data$mean_coverage,
max_iterations=100, 
change_threshold=0.01, 
epsilon=0)
}
