% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomplete_loglik.R
\name{incomplete_loglik}
\alias{incomplete_loglik}
\title{Function for calculating incomplete log likelihood}
\usage{
incomplete_loglik(
  xbeta,
  ff,
  firth = TRUE,
  outcome = outcome,
  epsilon = epsilon,
  covariate
)
}
\arguments{
\item{xbeta}{my fitted betas matrix multiplied by covariate}

\item{ff}{my estimated f's}

\item{firth}{use Firth penalty? default: TRUE}

\item{outcome}{length-n vector; this is the vector of a target gene's presence/absence; should be coded as 0 or 1}

\item{epsilon}{probability of observing a gene when it should be absent; probability between 0 and 1}

\item{covariate}{n x p matrix; this is the matrix for the primary predictor/covariate of interest}
}
\value{
incomplete log-likelihood value
}
\description{
Function for calculating incomplete log likelihood
}
\examples{
data("TM7_data")
# create design matrix
covariate <- model.matrix(~tongue, data = TM7_data) 
outcome <- TM7_data$`Ribosomal protein L27`
beta <- c(0.1,0.03)
my_xbeta <- covariate \%*\% beta
nn <- length(outcome) 
my_ff <- rep(mean(outcome), nn)

my_incomplete_LL <- incomplete_loglik(xbeta = my_xbeta, 
ff = my_ff, 
firth = TRUE, 
outcome = outcome, 
epsilon = 0, 
covariate = covariate)

}
