% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TM7_data}
\alias{TM7_data}
\title{Dataset of Saccharibacteria MAGs from Shaiber et al. 2020.}
\format{
A data frame with 43 rows and 732 variables: 
 \describe{
         \item{MAGs_Id}{Unique identifiers of each Saccharibacteria MAG}
         \item{site}{site of sample retrieval: tongue or plaque}
         \item{total_length}{total length of the MAG}
         \item{mean_coverage}{mean coverage of the MAG} 
         \item{tongue}{binary re-coding of sample origin site where 1 = tongue, 0 = plaque}
          }
          
          
 @source {This data was taken from Shaiber et al. 2020. https://doi.org/10.1186/s13059-020-02195-w}
 

 @examples
 data(TM7_data)
}
\usage{
TM7_data
}
\description{
This data contains 43 MAGs that have been made publicly available by Shaiber et al. 2020 at the link below.
}
\keyword{datasets}
